fix flaws
