+++ src/app/app.component.html
@@ -1,5 +1,1 @@
-<div class="loading-container" *ngIf="appLoading">
-  <mat-spinner></mat-spinner>
-</div>
-
 <router-outlet></router-outlet>
